chore: move SwapQuotesContext
